Remove content_policy_instrument and page_manager
These were left over from FourthParty and were never integrated with
OpenWPM's other instrumentation. The new context we've added to requests
and JS calls removes the need for the tricks done in page_manager. The
request context pulled in the content_policy_instrument was made
redundant by the context we pulled directly from the channel in the
http_instrument. Other context available in the content policy listener
can be pulled off of the channel's loadInfo property.