Refactor data handling to use a generic format internally.
The platform currently uses SQL throughout all of the instrumentation.
This makes it impossible to move to provide the option of an alternate
data format. This commit moves all SQL logic into a SQLiteAggregator and
uses simple dictionaries for passing data through the rest of the
platform. The BaseAggregator and BaseListener classes can be used to
support alternate output formats.