Handle raw HTTP POST data as well
In case that the HTTP POST data cannot be parsed as formData, it is also possible to access it as raw binary data. This was previously not supported by the extension and no HTTP POST data was logged for HTTP POST requests that failed to parse as formData by Firefox. This commit handles those cases as well.

The data is made available by Firefox as a list of 2-touples of filename - content pairs. The content is presented in a binary form. Since that content might not represent a valid utf8 string, a base64 encoding is applied so that the content is finally exported as a JSON encoded string.

This requires that the receiving code expects a new field named *post_body_raw*.