Fix test_js_profile_cookies
Previously, test_js_profile_cookies used _dump_profile_cookies_ which dumps the Firefox internal SQLite database to a table. This doesn't work very well anymore in the current version of Firefox and OpenWPM. Instead, the new extension provides a _javascript_cookies_ table that is filled with the cookies when _cookie_instrument_ is set to true. This commit updates the test so that the new table is used now and checks whether a cookie set by JavaScript is correctly recorded there.