Removing test for python packages.
Now that we have proper support for pinned packages, this is more
trouble than it's worth. Over the past few years the only time I've seen
this test fail is when I needed to track down a mismatch between a
package name on pip and the import name. That's only going to get worse
now that we are pinning all sub-dependencies.